Added sanity check to changeset handling
There is a total limit to how many parts you can append to a single batch. That limit is set to 100 at the time of writing, so we need to check for this when adding jobs to the batch. Added sanity checks when there is no changeset defined, to avoid throwing undefined errors.